refactor(columnar): extract TDD plan analysis
