+++ tests/meson.build
@@ -826,6 +826,7 @@ backend_src = files(
   '../wirelog/columnar/eval_dedup.c',
   '../wirelog/columnar/eval_delta.c',
   '../wirelog/columnar/eval_tdd_queue.c',
+  '../wirelog/columnar/eval_tdd_plan.c',
   '../wirelog/columnar/eval_plan.c',
   '../wirelog/columnar/eval_stack.c',
   '../wirelog/columnar/eval.c',
